Separate triedb utilities from monad-rpc into crate
