+++ monad-rpc/src/block_handlers.rs
@@ -1,6 +1,7 @@
 use alloy_primitives::aliases::{U256, U64};
 use log::{debug, trace};
 use monad_blockdb::{BlockTableKey, BlockValue};
+use monad_triedb_utils::{TriedbEnv, TriedbResult};
 use reth_primitives::BlockHash;
 use reth_rpc_types::{Block, BlockTransactions, Header, TransactionReceipt, Withdrawal};
 use serde::Deserialize;
@@ -13,7 +14,7 @@ use crate::{
     },
     eth_txn_handlers::{parse_tx_content, parse_tx_receipt},
     jsonrpc::JsonRpcError,
-    triedb::{decode_receipt, ReceiptDetails, TriedbEnv, TriedbResult},
+    receipt::{decode_receipt, ReceiptDetails},
 };
 
 fn parse_block_content(value: &BlockValue, return_full_txns: bool) -> Option<Block> {
