+++ monad-rpc/src/eth_json_types.rs
@@ -95,6 +95,15 @@ pub enum BlockTags {
     Default(BlockTagKey),
 }
 
+impl From<BlockTags> for monad_triedb_utils::BlockTags {
+    fn from(t: BlockTags) -> Self {
+        match t {
+            BlockTags::Number(q) => monad_triedb_utils::BlockTags::Number(q.0),
+            BlockTags::Default(k) => monad_triedb_utils::BlockTags::Default(k),
+        }
+    }
+}
+
 impl Default for BlockTags {
     fn default() -> Self {
         BlockTags::Default(BlockTagKey::Latest)
