+++ monad-triedb-utils/src/lib.rs
@@ -3,15 +3,19 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use alloy_primitives::{keccak256, Address, Bloom, Bytes};
-use alloy_rlp::{Decodable, Encodable, RlpDecodable, RlpEncodable};
+use alloy_primitives::keccak256;
+use alloy_rlp::{Decodable, Encodable};
 use log::debug;
 use monad_blockdb::BlockTagKey;
 use monad_triedb::Handle;
-use reth_primitives::{B256, U64, U8};
-use serde::{Deserialize, Serialize};
 
-use crate::eth_json_types::{BlockTags, EthAddress, EthStorageKey};
+pub enum BlockTags {
+    Number(u64),
+    Default(BlockTagKey),
+}
+
+pub type EthAddress = [u8; 20];
+pub type EthStorageKey = [u8; 32];
 
 #[derive(Clone)]
 pub struct TriedbEnv {
@@ -28,49 +32,6 @@ pub enum TriedbResult {
     Code(Vec<u8>),
     Receipt(Vec<u8>),
     BlockNum(u64),
-}
-
-#[derive(Debug, Clone, RlpEncodable, RlpDecodable, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReceiptDetails {
-    #[rlp(skip)]
-    #[rlp(default)]
-    pub tx_type: U8,
-    pub status: U64,
-    pub cumulative_gas_used: U64,
-    pub logs_bloom: Bloom,
-    pub logs: Vec<ReceiptLog>,
-}
-
-#[derive(Debug, Clone, RlpDecodable, RlpEncodable, Serialize, Deserialize)]
-pub struct ReceiptLog {
-    pub address: Address,
-    pub topics: Vec<B256>,
-    pub data: Bytes,
-}
-
-pub fn decode_receipt(rlp_buf: &mut &[u8]) -> alloy_rlp::Result<ReceiptDetails> {
-    let tx_type = decode_tx_type(rlp_buf)?;
-    let receipt = ReceiptDetails::decode(rlp_buf)?;
-    Ok(ReceiptDetails { tx_type, ..receipt })
-}
-
-pub fn decode_tx_type(rlp_buf: &mut &[u8]) -> Result<U8, alloy_rlp::Error> {
-    match rlp_buf.first() {
-        None => Err(alloy_rlp::Error::InputTooShort),
-        Some(&x) if x < 0xc0 => {
-            // first byte represents transaction type
-            let tx_type = match x {
-                1 => 1, // EIP2930
-                2 => 2, // EIP1559
-                // TODO: add support for EIP4844
-                _ => return Err(alloy_rlp::Error::Custom("InvalidTxnType")),
-            };
-            *rlp_buf = &rlp_buf[1..]; // advance the buffer
-            Ok(U8::from(tx_type))
-        }
-        Some(_) => Ok(U8::from(0)), // legacy transactions do not have first byte as transaction type
-    }
 }
 
 impl TriedbEnv {
@@ -112,7 +73,7 @@ impl TriedbEnv {
 
                 // parse block tag
                 let block_num = match block_tag {
-                    BlockTags::Number(q) => q.0,
+                    BlockTags::Number(q) => q,
                     BlockTags::Default(t) => match t {
                         BlockTagKey::Latest => TriedbEnv::latest_block(db),
                         BlockTagKey::Finalized => TriedbEnv::latest_block(db),
@@ -185,7 +146,7 @@ impl TriedbEnv {
 
                 // parse block tag
                 let block_num = match block_tag {
-                    BlockTags::Number(q) => q.0,
+                    BlockTags::Number(q) => q,
                     BlockTags::Default(t) => match t {
                         BlockTagKey::Latest => TriedbEnv::latest_block(db),
                         BlockTagKey::Finalized => TriedbEnv::latest_block(db),
@@ -226,7 +187,7 @@ impl TriedbEnv {
 
                 // parse block tag
                 let block_num = match block_tag {
-                    BlockTags::Number(q) => q.0,
+                    BlockTags::Number(q) => q,
                     BlockTags::Default(t) => match t {
                         BlockTagKey::Latest => TriedbEnv::latest_block(db),
                         BlockTagKey::Finalized => TriedbEnv::latest_block(db),
@@ -286,7 +247,7 @@ impl TriedbEnv {
 
         key_nibbles.push(state_nibble);
 
-        let hashed_addr = keccak256(addr.0);
+        let hashed_addr = keccak256(addr);
         for byte in &hashed_addr {
             key_nibbles.push(*byte >> 4);
             key_nibbles.push(*byte & 0xF);
@@ -310,13 +271,13 @@ impl TriedbEnv {
         let state_nibble = 0_u8;
         key_nibbles.push(state_nibble);
 
-        let hashed_addr = keccak256(addr.0);
+        let hashed_addr = keccak256(addr);
         for byte in &hashed_addr {
             key_nibbles.push(*byte >> 4);
             key_nibbles.push(*byte & 0xF);
         }
 
-        let hashed_at = keccak256(at.0);
+        let hashed_at = keccak256(at);
         for byte in &hashed_at {
             key_nibbles.push(*byte >> 4);
             key_nibbles.push(*byte & 0xF);
@@ -385,20 +346,5 @@ impl TriedbEnv {
             .collect();
 
         (key, num_nibbles)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use reth_primitives::U8;
-
-    use crate::hex;
-
-    #[test]
-    fn decode_receipt() {
-        let rlp = hex::decode("0x02f90109018301ab80b9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c0").unwrap();
-        let mut rlp_buf = rlp.as_slice();
-        let receipt = super::decode_receipt(&mut rlp_buf).unwrap();
-        assert_eq!(receipt.tx_type, U8::from(2));
     }
 }
